// react-app/src/misc/openapi.d.ts
export interface components {
    schemas: {
// ...
        CreateNoteRequest: {
            section: string;
            book?: string | null;
        };
// ...
        NoteItemResponse: {
            /** Format: int64 */
            id?: number;
            createdAt?: string | null;
            updatedAt?: string | null;
            title?: string | null;
            section?: string | null;
            content?: string | null;
            preview?: string | null;
            book?: string | null;
            tags?: string[] | null;
            sources?: components["schemas"]["NoteSource"][] | null;
        };
// ...
        SearchNotesRequest: {
            section?: string | null;
            searchTerm?: string | null;
            /** Format: int32 */
            page?: number;
            book?: string | null;
            tags?: string[] | null;
            inRandomOrder?: boolean;
            withoutBook?: boolean;
            withoutTags?: boolean;
            /** Format: date */
            fromDate?: string | null;
            /** Format: date */
            toDate?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}